+++ primevue/custompt.ts
@@ -1,7 +1,0 @@
-const MyCustomPT = {
-  button: {
-      size: 'small'
-  },
-}
-
-export default MyCustomPT;