+++ CHANGELOG.md
@@ -1,6 +1,67 @@
 # Changelog
 
 
+## v1.3.0
+
+[compare changes](https://github.com/aatrooox/blog.zzao.club/compare/v1.2.2...v1.3.0)
+
+### 🚀 Enhancements
+
+- 评论时增加邮件提示 ([f869c13](https://github.com/aatrooox/blog.zzao.club/commit/f869c13))
+
+### ❤️ Contributors
+
+- Aatrox <gnakzz@qq.com>
+
+## v1.2.2
+
+[compare changes](https://github.com/aatrooox/blog.zzao.club/compare/v1.2.1...v1.2.2)
+
+### 🏡 Chore
+
+- Ci ([518f643](https://github.com/aatrooox/blog.zzao.club/commit/518f643))
+
+### ❤️ Contributors
+
+- Aatrox <gnakzz@qq.com>
+
+## v1.2.1
+
+[compare changes](https://github.com/aatrooox/blog.zzao.club/compare/v1.2.0...v1.2.1)
+
+### 🩹 Fixes
+
+- Settings prerender false ([1f5f7d5](https://github.com/aatrooox/blog.zzao.club/commit/1f5f7d5))
+
+### ❤️ Contributors
+
+- Aatrox <gnakzz@qq.com>
+
+## v1.2.0
+
+[compare changes](https://github.com/aatrooox/blog.zzao.club/compare/v1.1.1...v1.2.0)
+
+### 🚀 Enhancements
+
+- 新增邮件通知接口 ([8566e1f](https://github.com/aatrooox/blog.zzao.club/commit/8566e1f))
+- 增加用户配置表 ([ddbfc3e](https://github.com/aatrooox/blog.zzao.club/commit/ddbfc3e))
+- ⚠️  新增用户设置界面; 组件位置移动; 优化代码 ([2191ecf](https://github.com/aatrooox/blog.zzao.club/commit/2191ecf))
+
+### 🏡 Chore
+
+- 优化 ([48feb6c](https://github.com/aatrooox/blog.zzao.club/commit/48feb6c))
+- Gitea env ([012b488](https://github.com/aatrooox/blog.zzao.club/commit/012b488))
+- Env example ([8f19fb6](https://github.com/aatrooox/blog.zzao.club/commit/8f19fb6))
+- Nuxt/content config; pm2 json config; readme ([532a247](https://github.com/aatrooox/blog.zzao.club/commit/532a247))
+
+#### ⚠️ Breaking Changes
+
+- ⚠️  新增用户设置界面; 组件位置移动; 优化代码 ([2191ecf](https://github.com/aatrooox/blog.zzao.club/commit/2191ecf))
+
+### ❤️ Contributors
+
+- Aatrox <gnakzz@qq.com>
+
 ## v1.1.1
 
 [compare changes](https://github.com/aatrooox/blog.zzao.club/compare/v1.1.0...v1.1.1)